Reformat Code with Code formatter
Changes are whitespace, indentation and position of braces (brace position according to One True Brace Style variant of the Kernighan and Ritchie style). Also the substitution of aliases with the original, real name, and the correction of case to match the official names and Coding Recommendations.

Done with the VSCode Powershell extension Code formatter.